pyproject.toml
[project]
name = "etl"
version = "0.1.0"
description = ""
authors = [
    {name = "Dmytro",email = "zandarn@ukr.net"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "alembic (==1.15.2)",
    "annotated-types (==0.7.0)",
    "anyio (==3.7.1)",
    "async-timeout (==5.0.1)",
    "backports-tarfile (==1.2.0)",
    "black (==25.1.0)",
    "build (==1.2.2.post1)",
    "cachecontrol (==0.14.3)",
    "certifi (==2025.4.26)",
    "charset-normalizer (==3.4.2)",
    "click (==8.1.8)",
    "colorama (==0.4.6)",
    "coverage (==7.8.0)",
    "distlib (==0.3.9)",
    "dnspython (==2.7.0)",
    "dulwich (==0.22.8)",
    "email-validator (==2.2.0)",
    "exceptiongroup (==1.2.2)",
    "fastapi (==0.115.12)",
    "fastapi-cli (==0.0.7)",
    "fastjsonschema (==2.21.1)",
    "filelock (==3.18.0)",
    "findpython (==0.6.3)",
    "greenlet (==3.2.1)",
    "h11 (==0.16.0)",
    "httpcore (==1.0.9)",
    "httptools (==0.6.4)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "importlib-metadata (==8.7.0)",
    "iniconfig (==2.1.0)",
    "installer (==0.7.0)",
    "jaraco-classes (==3.4.0)",
    "jaraco-context (==6.0.1)",
    "jaraco-functools (==4.1.0)",
    "jinja2 (==3.1.6)",
    "keyring (==25.6.0)",
    "mako (==1.3.10)",
    "markdown-it-py (==3.0.0)",
    "markupsafe (==3.0.2)",
    "mdurl (==0.1.2)",
    "more-itertools (==10.7.0)",
    "msgpack (==1.1.0)",
    "mypy-extensions (==1.1.0)",
    "mysqlclient (==2.2.7)",
    "packaging (==25.0)",
    "pathspec (==0.12.1)",
    "pbs-installer (==2025.6.6)",
    "pkginfo (==1.12.1.2)",
    "platformdirs (==4.3.8)",
    "pluggy (==1.6.0)",
    "pydantic (==2.11.4)",
    "pydantic-settings (==2.9.1)",
    "pydantic-core (==2.33.2)",
    "pygments (==2.19.1)",
    "pyproject-hooks (==1.2.0)",
    "pytest (==8.3.5)",
    "pytest-cov (==6.1.1)",
    "python-dotenv (==1.1.0)",
    "python-multipart (==0.0.20)",
    "pywin32-ctypes (==0.2.3)",
    "pyyaml (==6.0.2)",
    "rapidfuzz (==3.13.0)",
    "redis (==6.2.0)",
    "requests (==2.32.3)",
    "requests-toolbelt (==1.0.0)",
    "rich (==14.0.0)",
    "rich-toolkit (==0.14.4)",
    "shellingham (==1.5.4)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.40)",
    "sqlmodel (==0.0.24)",
    "starlette (==0.46.2)",
    "tomli (==2.2.1)",
    "tomlkit (==0.13.3)",
    "trove-classifiers (==2025.5.9.12)",
    "typer (==0.15.3)",
    "typing-inspection (==0.4.0)",
    "typing-extensions (==4.13.2)",
    "urllib3 (==2.4.0)",
    "uvicorn (==0.34.2)",
    "virtualenv (==20.31.2)",
    "watchfiles (==1.0.5)",
    "watchgod (==0.8.2)",
    "websockets (==15.0.1)",
    "zipp (==3.23.0)",
    "zstandard (==0.23.0)",
    "cleo (==2.1.0)",
    "crashtest (==0.4.1)",
    "poetry (==2.1.3)",
    "poetry-core (==2.1.3)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
